Moved out ITextCompletionExtension to its own file
